server/bigquery_loader/requirements.txt
google-cloud-pubsub
google-cloud-bigquery
flask
gunicorn
